feature(eapi6): initial EAPI 6 profile and function stubs
We should figure out a way to dynamically build a list of export blocks
for EAPI dependent functions, i.e. we shouldn't block exports from
eclasses for similarly named functions but instead override them with
internal PM implementations when EAPI provided functionality comes into
existence.

The current method involves explicitly adding the related functions to
the INTERNAL_FUNCS list the only current affected function being "usex"
that is defined in EAPI 5 and in the eutils.eclass for EAPI < 5. This
requires us to track implementations within eclasses which can be
problematic.